changed contactNumber to Long datatype

+++ src/main/java/com/sdp/sports_management/bean/User.java
@@ -35,7 +35,7 @@ public class User {
     private String emailId;
 
     @Column(length=10)
-    private Integer contactNumber;
+    private Long contactNumber;
 
     @Column(length=50,nullable = false)
     private String password;
@@ -71,11 +71,11 @@ public class User {
 
 
 
-    public Integer getContactNumber() {
+    public Long getContactNumber() {
         return contactNumber;
     }
 
-    public void setContactNumber(Integer contactNumber) {
+    public void setContactNumber(Long contactNumber) {
         this.contactNumber = contactNumber;
     }
 
@@ -95,7 +95,7 @@ public class User {
         isAdminFlag = adminFlag;
     }
 
-    public User(String firstName, String lastName, String emailId, Integer contactNumber, String password, Boolean isAdminFlag) {
+    public User(String firstName, String lastName, String emailId, Long contactNumber, String password, Boolean isAdminFlag) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.emailId = emailId;
